+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 - Include `pixelscript_m.h` in `pixelscript_cpp.hpp`
 - Free the error message in `consume_error`.
 - Add `HWrapper` to `pixelscript_cpp.hpp` wrapper.
+- Added `PXSVariant` to `pixelscript_cpp.hpp` wrapper.